Fix db relation
