+++ my_project/src/Controller/HomepageController.php
@@ -36,26 +36,6 @@ class HomepageController extends AbstractController
 
         $searchTerm = $request->query->get('search');
 
-        // TODO: To remove once data properly obtained
-        $hardcodedTestChars =  [
-            ['id' => 1, 'name' => 'Luck', 'image' => 'image.png', 'mass' => '100', 'height' => '185'],
-            ['id' => 2, 'name' => 'Obi Wan', 'image' => 'image2.png', 'mass' => '80', 'height' => '180'],
-            ['id' => 3, 'name' => 'Doku', 'image' => 'image3.png', 'mass' => '90', 'height' => '190']
-        ];
-        if ($searchTerm) {
-            $filteredChars = array_filter($hardcodedTestChars, function($char) use ($searchTerm) {
-                return stripos($char['name'], $searchTerm) !== false;
-            });
-            $finalCharacters = array_values($filteredChars);
-        } else {
-            $finalCharacters = $hardcodedTestChars;
-        }
-        return $this->render('homepage/index.html.twig', [
-            'characters' => $finalCharacters,
-            'searchTerm' => $searchTerm,
-        ]);
-        // TODO: END TODO
-
         if ($searchTerm) {
             $characters = $this->entityManager->getRepository(Character::class)
                 ->findBy(['name' => $searchTerm]);
